requirements.txt
requests
beautifulsoup4
tabula-py
camelot-py
opencv-python
ghostscript
jpype1
